合并player和npc的状态机
