+++ Entities/Player/player.tscn
@@ -16,11 +16,11 @@
 [ext_resource type="PackedScene" uid="uid://co8hoc0oc4bbb" path="res://Entities/shadow.tscn" id="15_oprun"]
 [ext_resource type="Script" uid="uid://dhk31wxarmn7" path="res://Components/KnockbackComponent.cs" id="16_qfm1y"]
 [ext_resource type="Script" uid="uid://e760loqgo1xv" path="res://Core/StateMachine.cs" id="18_a38lo"]
-[ext_resource type="Script" uid="uid://d1h1usnwy8f5q" path="res://Entities/Player/States/RunState.cs" id="19_4ni07"]
-[ext_resource type="Script" uid="uid://bviqnfqsuhf8o" path="res://Entities/Player/States/IdleState.cs" id="20_l71n6"]
-[ext_resource type="Script" uid="uid://dj8mj2ea8x5u4" path="res://Entities/Player/States/DeadState.cs" id="21_ke2ow"]
-[ext_resource type="Script" uid="uid://cxffbk7dra3y7" path="res://Entities/Player/States/StaggerState.cs" id="22_ujl30"]
-[ext_resource type="Script" uid="uid://cux04vlwjftrw" path="res://Entities/Player/States/AttackState.cs" id="23_4ni07"]
+[ext_resource type="Script" path="res://Entities/States/MoveState.cs" id="19_4ni07"]
+[ext_resource type="Script" path="res://Entities/States/IdleState.cs" id="20_l71n6"]
+[ext_resource type="Script" path="res://Entities/States/DeadState.cs" id="21_ke2ow"]
+[ext_resource type="Script" path="res://Entities/States/StaggerState.cs" id="22_ujl30"]
+[ext_resource type="Script" path="res://Entities/States/AttackState.cs" id="23_4ni07"]
 [ext_resource type="Script" uid="uid://ckedmoevht7w4" path="res://Components/SelectionManager.cs" id="24_u8v8j"]
 
 [sub_resource type="CapsuleShape2D" id="CapsuleShape2D_tuyoq"]
@@ -430,7 +430,7 @@ scale = Vector2(1, 0.5)
 [node name="StateMachine" type="Node" parent="." unique_id=548797990]
 script = ExtResource("18_a38lo")
 
-[node name="Run" type="Node" parent="StateMachine" unique_id=1258306878]
+[node name="Move" type="Node" parent="StateMachine" unique_id=1258306878]
 script = ExtResource("19_4ni07")
 
 [node name="Idle" type="Node" parent="StateMachine" unique_id=1812659606]
@@ -441,6 +441,9 @@ script = ExtResource("21_ke2ow")
 
 [node name="Stagger" type="Node" parent="StateMachine" unique_id=1936291366]
 script = ExtResource("22_ujl30")
+ExitPolicy = 0
+StaggerDuration = 0.25
+ClearMoveIntentOnEnter = true
 
 [node name="Attack" type="Node" parent="StateMachine" unique_id=759954878]
 script = ExtResource("23_4ni07")
