+++ Entities/States/DeadState.cs
@@ -1,13 +1,9 @@
-namespace AlongJourney.Entities.Player.States;
+namespace AlongJourney.Entities.States;
 
 using Godot;
-using AlongJourney.Entities;
 using AlongJourney.Core;
+using AlongJourney.Entities;
 
-/// <summary>
-/// 死亡状态：Actor 死亡时的状态
-/// 死亡后不再处理任何逻辑
-/// </summary>
 public partial class DeadState : State
 {
     public override void Enter()
@@ -17,19 +13,16 @@ public partial class DeadState : State
             return;
         }
 
-        // 停止所有移动
         Owner.Velocity = Vector2.Zero;
         Owner.SetBlackboardValue(Actor.BlackboardKeys.MoveDirection, Vector2.Zero);
         Owner.SetBlackboardValue(Actor.BlackboardKeys.InputVector, Vector2.Zero);
         Owner.SetBlackboardValue(Actor.BlackboardKeys.IsDead, true);
-        
-        // 禁用碰撞和受击盒
+
         Owner.SetCollisionEnabled(false);
         Owner.SetHurtboxEnabled(false);
     }
 
     public override void Update(double delta)
     {
-        // 死亡状态不处理任何逻辑
     }
 }
